Replace table reference with indexed table
`TableReference` was created to implement index hints. But I was also aiming to
use it for sub queries in the future. However, as I am working on sub query now,
it turns out not to be the best design for sub queries, and this interface
introduces unnecessary complications into `GenericQuery`. So it is removed and
replaced.

Users should not be affected in most cases. The `AbstractTable#with` still works
in `GenericQuery`.