Handle throwables while executing transactions (#245)
Both methods which execute transactions (query and execute) only catch Exceptions, not Throwables. As a result, if a Throwable is thrown in the middle of a transaction, then we only execute the finally block which returns the connection (and we don't explicitly rollback the transaction as we would have done inside the catch(Exception e) block. The unfortunate side effect of this behavior is that it commits partially-completed transactions; returning a connection calls .setAutoCommit(true) which commits any commands built up during the uncommitted transaction which leaves us with inconsistent state.

Changes introduced in this PR: 

1. remove the finally block, and instead return the connection at the end of both the try and catch(Exception e) blocks.
2. add a catch(Throwable t) block in which we try to rollback any uncommitted changes, but do not attempt to return the connection.
3. In the interest of DRY, refactor execute to be a wrapper on top of query instead of an almost exact replica.