Ensure that pooled connections are tested before being used.
This is technically covered by the implementation of most methods in
`AbstractDatabaseModel`, which check for exceptions and then call
`resetConnection`, but this explicitly ensures that no matter the
underlying method implementation, we'll still validate the connection
before using it. We also validate while connections are idle to reduce
the number of destroy/create cycles that we might need at the time of
borrowing a connection. This is especially relevant for long running
processes that might attempt to use very old pooled connections.