Fix connection leak caused by exceptions during rollback. (#258)
* Fix connection leak caused by exceptions during rollback.

If an exception is thrown while attempting to rollback a transaction,
then the connection is never returned to the connection pool. This
slowly leaks connections until there are none left.

* Introduce `IDbManager#invalidateConnection` and use it on rollback failure.

This avoids the confusion of returning a broken connection or one that
may have a partially finished transaction that should not commit be
committed, which returning it would do.