Include milliseconds in SQL statements
When building SQL statements for DATETIME and TIMESTAMP columns, include
milliseconds (fractional seconds precision of 3).

Jack represents DATETIME and TIMESTAMP types as ms held in longs, so it
is convenient to support the same level of precision when building SQL
statements.

It's been confirmed that althought MySQL 5.5 does not support storing
fractional seconds in time values, it can parse them; they'll be
truncated in the tables.

Jira: INGEST-2133