Don't NPE when config option is not specified
When some of the options are incorrectly specified, jack will
throw an NPE when trying to craft a user-friendly error message.

You can see this in the current version by not specifying the
`adaptor` parameter.

This change attempts to avoid such an error by not failing when
crafting an exception error message.